Se modficaron las mayorias de import por que se cambiaron los nombres
