+++ changuitasAppFront/Pantallas/RecuperarAcceso/RecuperarContrasena1.tsx
@@ -4,7 +4,7 @@ import { Ionicons } from "@expo/vector-icons";
 import { useNavigation, NavigationProp } from '@react-navigation/native';
 import { Snackbar } from "react-native-paper";
 import { RootStackParamList } from '../../navegacion/AppNavigator';
-import API_URL from "../../auxiliares/API_URL";
+import API_URL from "../../utils/API_URL";
 import EstilosRecuperarContrasena1 from "../RecuperarAcceso/estilos/EstilosRecuperarContrasena1";
 import { NavBarSuperior } from "../../componentes/NavBarSuperior";
 import { Platform, useWindowDimensions } from 'react-native';
