+++ changuitasAppFront/Pantallas/Registro-Login/estilos/EstilosPantallaAyuda.js
@@ -3,7 +3,9 @@ import { StyleSheet } from 'react-native';
 const EstilosPantallaAyuda = StyleSheet.create({
   contenedor: {
     flex: 1,
-    backgroundColor: '#191a2e',
+  },
+  degradado: {
+    flex: 1,
   },
   encabezado: {
     flexDirection: 'row',
@@ -12,27 +14,23 @@ const EstilosPantallaAyuda = StyleSheet.create({
     paddingHorizontal: 20,
     paddingTop: 10,
     paddingBottom: 15,
-    backgroundColor: '#2a2b45',
     marginTop: 50,
   },
   textoInicio: {
     fontSize: 24,
     color: '#F2F2F2',
     fontWeight: '600',
-    color:"white",
   },
   menuPuntos: {
     fontSize: 24,
     color: '#F2F2F2',
     fontWeight: '600',
-    color:"#fd6b2f"
   },
   tituloAyuda: {
     fontSize: 18,
-    color: 'white',
+    color: '#FC6A30',
     textAlign: 'center',
     marginVertical: 10,
-    fontWeight:"bold",
   },
   datosPreguntas: {
     paddingHorizontal: 20,
@@ -40,13 +38,12 @@ const EstilosPantallaAyuda = StyleSheet.create({
   },
   tituloPregunta: {
     fontSize: 16,
-    color: '#fd6b2f',
+    color: '#FC6A30',
     marginBottom: 5,
-    fontWeight:"bold",
   },
   infoPregunta: {
     fontSize: 16,
-    color: 'white',
+    color: '#B3B3B3',
     marginBottom: 5,
   },
   contenidoResponsivo: {
