+++ changuitasAppFront/Pantallas/Servicios/AgregarServicio1.tsx
@@ -4,10 +4,10 @@ import { Checkbox } from 'react-native-paper';
 import { Ionicons } from "@expo/vector-icons";
 import { useNavigation, NavigationProp } from '@react-navigation/native';
 import { RootStackParamList } from '../../navegacion/AppNavigator';
-import BarraNavegacionInferior from '../../auxiliares/BarraNavegacionInferior';
+import BarraNavegacionInferior from '../../utils/BarraNavegacionInferior';
 import EstilosAgregarServicio1 from './estilos/EstilosAgregarServicio1';
 import EncabezadoPerfil from '../../componentes/perfilesUsuarios/EncabezadoPerfil';
-import MenuDesplegable from '../../auxiliares/MenuDesplegable';
+import MenuDesplegable from '../../componentes/MenuDesplegable';
 import { AuthContext } from '../../autenticacion/auth';
 import {cerrarSesion} from '../../autenticacion/authService';
 import { Button } from '../../componentes/Buttons';
