+++ changuitasAppFront/Pantallas/Servicios/AgregarServicio2.tsx
@@ -4,9 +4,9 @@ import { Ionicons } from "@expo/vector-icons";
 import { useNavigation, useRoute, RouteProp, NavigationProp } from '@react-navigation/native';
 import { RootStackParamList } from '../../navegacion/AppNavigator';
 import { Picker } from '@react-native-picker/picker';
-import BarraNavegacionInferior from '../../auxiliares/BarraNavegacionInferior';
+import BarraNavegacionInferior from '../../utils/BarraNavegacionInferior';
 import EstilosAgregarServicio2 from './estilos/EstilosAgregarServicio2';
-import API_URL from '../../auxiliares/API_URL';
+import API_URL from '../../utils/API_URL';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import CampoDescripcion from '../../componentes/inputs/CampoDescripcion';
 import { Button } from '../../componentes/Buttons';
