+++ changuitasAppFront/Pantallas/Usuario/EditarPerfil.tsx
@@ -4,18 +4,18 @@ import { Ionicons } from "@expo/vector-icons";
 import { useNavigation, NavigationProp } from '@react-navigation/native';
 import { RootStackParamList } from '../../navegacion/AppNavigator';
 import AsyncStorage from '@react-native-async-storage/async-storage';
-import API_URL from '../../auxiliares/API_URL';;
+import API_URL from '../../utils/API_URL';;
 import axios from 'axios';
 import { Snackbar } from 'react-native-paper';
 import EstilosEditarPerfil from './estilos/EstilosEditarPerfil';
 import {cerrarSesion} from '../../autenticacion/authService';
 import { AuthContext } from '../../autenticacion/auth';
-import { mostrarOpcionesSelectorImagen } from '../../auxiliares/seleccionImagen';
-import BarraPestanasPerfil from '../../auxiliares/BarraPestanasPerfil';
-import BarraNavegacionInferior from '../../auxiliares/BarraNavegacionInferior';
-import { ImageCropperWeb } from '../../auxiliares/ImageCropperWeb';
+import { mostrarOpcionesSelectorImagen } from '../../utils/seleccionImagen';
+import BarraPestanasPerfil from '../../utils/BarraPestanasPerfil';
+import BarraNavegacionInferior from '../../utils/BarraNavegacionInferior';
+import { ImageCropperWeb } from '../../componentes/ImageCropperWeb';
 import { guardarCambios } from './auxiliar/guardarCambios';
-import MenuDesplegable from '../../auxiliares/MenuDesplegable';
+import MenuDesplegable from '../../componentes/MenuDesplegable';
 import CustomModal from '../../componentes/CustomModal';
 import estilosModal from '../../componentes/estilosModal';
 import EncabezadoPerfil from '../../componentes/perfilesUsuarios/EncabezadoPerfil';
