+++ changuitasAppFront/Pantallas/Usuario/PerfilUsuario.tsx
@@ -8,11 +8,11 @@ import { TouchableWithoutFeedback} from 'react-native';
 import { RootStackParamList } from '../../navegacion/AppNavigator';
 import {cerrarSesion} from '../../autenticacion/authService';
 import EstilosPerfilUsuario from './estilos/EstilosPerfilUsuario';
-import BarraNavegacionInferior from '../../auxiliares/BarraNavegacionInferior';
-import API_URL from '../../auxiliares/API_URL';
-import BarraPestanasPerfil from '../../auxiliares/BarraPestanasPerfil';
+import BarraNavegacionInferior from '../../utils/BarraNavegacionInferior';
+import API_URL from '../../utils/API_URL';
+import BarraPestanasPerfil from '../../utils/BarraPestanasPerfil';
 import { AuthContext } from '../../autenticacion/auth';
-import MenuDesplegable from '../../auxiliares/MenuDesplegable';
+import MenuDesplegable from '../../componentes/MenuDesplegable';
 import CustomSnackbar from '../../componentes/CustomSnackbar';
 import EncabezadoPerfil from '../../componentes/perfilesUsuarios/EncabezadoPerfil';
 import ImagenPerfilUsuario from '../../componentes/perfilesUsuarios/ImagenPerfilUsuario';
